jobs: mon-web-package now takes an argument.

+++ jobs/web/mon-web-package.sh
@@ -3,8 +3,15 @@
 set -e
 set -x
 
+# Check arguments.
+if [ "$#" -lt 1 ] ; then
+  echo "USAGE: $0 <6|7>"
+  exit 1
+fi
+CENTOS_VERSION="$1"
+
 # Pull mon-build-dependencies container.
-docker pull ci.int.centreon.com:5000/mon-build-dependencies:centos6
+docker pull ci.int.centreon.com:5000/mon-build-dependencies:centos$CENTOS_VERSION
 
 # Create input and output directories for docker-rpm-builder.
 rm -rf input
@@ -87,7 +94,7 @@ cp packaging-centreon-web/rpm/centreon.spectemplate input/
 cp packaging-centreon-web/src/* input
 
 # Build RPMs.
-docker-rpm-builder dir ci.int.centreon.com:5000/mon-build-dependencies:centos6 input output
+docker-rpm-builder dir ci.int.centreon.com:5000/mon-build-dependencies:centos$CENTOS_VERSION input output
 
 # Copy files to server.
 CES_VERSION='3.0'
